pyproject.toml
[project]
name = "IFComunidade_backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "carloshenriqueok", email = "ifcarloshenrique@gmail.com"},
]
dependencies = ["django>=5.2.6", "djangorestframework>=3.16.1", "django-cors-headers>=4.7.0", "setuptools>=80.9.0", "djangorestframework-simplejwt>=5.5.1", "paramiko>=4.0.0", "Pillow>=11.3.0", "python-magic>=0.4.27; sys_platform == \"linux\"", "drf-spectacular>=0.28.0", "dj-database-url>=3.0.1", "python-dotenv>=1.1.1", "shell>=1.0.1", "psycopg[binary]>=3.2.10"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
